move some reusable diag code to ts/api
