+++ include/api/DbgCtl.h
@@ -25,6 +25,7 @@
 
 #include "api/SourceLocation.h"
 #include "api/ts_diag_levels.h"
+#include "swoc/BufferWriter.h"
 #include <atomic>
 #include <utility>
 #include <functional>
@@ -32,6 +33,8 @@
 #include <ts/apidefs.h> // For TS_PRINTFLIKE
 
 class DiagsConfigState;
+
+enum DiagsShowLocation { SHOW_LOCATION_NONE = 0, SHOW_LOCATION_DEBUG, SHOW_LOCATION_ALL };
 
 class DebugInterface
 {
@@ -44,6 +47,12 @@ public:
 
   static DebugInterface *get_instance();
   static void set_instance(DebugInterface *);
+
+  // Generate the default diagnostics format string for the given parameters.
+  // @return The offset in the format string of the timestamp (in case the caller doesn't want to include that)
+  static size_t generate_format_string(swoc::LocalBufferWriter<1024> &format_writer, const char *debug_tag, DiagsLevel diags_level,
+                                       const SourceLocation *loc, DiagsShowLocation show_location, const char *format_string);
+  static const char *level_name(DiagsLevel dl);
 };
 
 class DbgCtl
