+++ src/api/DbgCtl.cc
@@ -21,21 +21,27 @@
   limitations under the License.
  */
 
+#include <ctime>
 #include <mutex>
 #include <map>
 #include <cstring>
 #include <atomic>
 #include <cstdarg>
+#include <cinttypes>
 
 #include "api/SourceLocation.h"
 
+#include "api/ts_diag_levels.h"
+#include "api/ts_bw_format.h"
 #include "tscore/ink_assert.h"
 #include "api/DbgCtl.h"
 
+#include "tscore/ink_config.h"
+
+using namespace swoc::literals;
+
 DbgCtl::DbgCtl(DbgCtl &&src)
 {
-  ink_release_assert(src._ptr != &_No_tag_dummy());
-
   _ptr     = src._ptr;
   src._ptr = &_No_tag_dummy();
 }
@@ -43,8 +49,6 @@ DbgCtl::DbgCtl(DbgCtl &&src)
 DbgCtl &
 DbgCtl::operator=(DbgCtl &&src)
 {
-  ink_release_assert(&_No_tag_dummy() == _ptr);
-
   new (this) DbgCtl{std::move(src)};
 
   return *this;
@@ -211,12 +215,20 @@ void
 DbgCtl::print(char const *tag, char const *file, char const *function, int line, char const *fmt_str, ...)
 {
   DebugInterface *p = DebugInterface::get_instance();
-  ink_release_assert(p);
   SourceLocation src_loc{file, function, line};
-  va_list args;
-  va_start(args, fmt_str);
-  p->print_va(tag, DL_Diag, &src_loc, fmt_str, args);
-  va_end(args);
+  if (p) {
+    va_list args;
+    va_start(args, fmt_str);
+    p->print_va(tag, DL_Diag, &src_loc, fmt_str, args);
+    va_end(args);
+  } else {
+    swoc::LocalBufferWriter<1024> format_writer;
+    DebugInterface::generate_format_string(format_writer, tag, DL_Diag, &src_loc, SHOW_LOCATION_DEBUG, fmt_str);
+    va_list args;
+    va_start(args, fmt_str);
+    vprintf(format_writer.data(), args);
+    va_end(args);
+  }
 }
 
 std::atomic<int> DbgCtl::_config_mode{0};
@@ -225,14 +237,44 @@ bool
 DbgCtl::_override_global_on()
 {
   DebugInterface *p = DebugInterface::get_instance();
-  ink_release_assert(p);
-  return p->get_override();
+  if (p) {
+    return p->get_override();
+  } else {
+    return false;
+  }
 }
 
 namespace
 {
 static DebugInterface *di_inst;
+
+bool
+location(const SourceLocation *loc, DiagsShowLocation show, DiagsLevel level)
+{
+  if (loc && loc->valid()) {
+    switch (show) {
+    case SHOW_LOCATION_ALL:
+      return true;
+    case SHOW_LOCATION_DEBUG:
+      return level <= DL_Debug;
+    default:
+      return false;
+    }
+  }
+
+  return false;
 }
+
+//////////////////////////////////////////////////////////////////////////////
+//
+//      const char *Diags::level_name(DiagsLevel dl)
+//
+//      This routine returns a string name corresponding to the error
+//      level <dl>, suitable for us as an output log entry prefix.
+//
+//////////////////////////////////////////////////////////////////////////////
+
+} // namespace
 
 DebugInterface *
 DebugInterface::get_instance()
@@ -245,4 +287,109 @@ DebugInterface::set_instance(DebugInterface *i)
 {
   di_inst = i;
   DbgCtl::update([&](const char *t) { return i->debug_tag_activated(t); });
+}
+
+const char *
+DebugInterface::level_name(DiagsLevel dl)
+{
+  switch (dl) {
+  case DL_Diag:
+    return ("DIAG");
+  case DL_Debug:
+    return ("DEBUG");
+  case DL_Status:
+    return ("STATUS");
+  case DL_Note:
+    return ("NOTE");
+  case DL_Warning:
+    return ("WARNING");
+  case DL_Error:
+    return ("ERROR");
+  case DL_Fatal:
+    return ("FATAL");
+  case DL_Alert:
+    return ("ALERT");
+  case DL_Emergency:
+    return ("EMERGENCY");
+  default:
+    return ("DIAG");
+  }
+}
+
+namespace
+{
+
+struct DiagTimestamp {
+  std::chrono::time_point<std::chrono::system_clock> ts = std::chrono::system_clock::now();
+};
+
+swoc::BufferWriter &
+bwformat(swoc::BufferWriter &w, swoc::bwf::Spec const &spec, DiagTimestamp const &ts)
+{
+  auto epoch = std::chrono::system_clock::to_time_t(ts.ts);
+  swoc::LocalBufferWriter<48> lw;
+
+  ctime_r(&epoch, lw.aux_data());
+  lw.commit(19); // keep only leading text.
+  lw.print(".{:03}", std::chrono::time_point_cast<std::chrono::milliseconds>(ts.ts).time_since_epoch().count() % 1000);
+  w.write(lw.view().substr(4));
+
+  return w;
+}
+
+struct DiagThreadname {
+  char name[32];
+
+  DiagThreadname() {
+#if defined(HAVE_PTHREAD_GETNAME_NP)
+    pthread_getname_np(pthread_self(), name, sizeof(name));
+#elif defined(HAVE_PTHREAD_GET_NAME_NP)
+    pthread_get_name_np(pthread_self(), name, sizeof(name));
+#elif defined(HAVE_SYS_PRCTL_H) && defined(PR_GET_NAME)
+    prctl(PR_GET_NAME, name, 0, 0, 0);
+#else
+    snprintf(name, sizeof(name), "0x%" PRIx64, (uint64_t)pthread_self());
+#endif
+  }
+};
+
+swoc::BufferWriter &
+bwformat(swoc::BufferWriter &w, swoc::bwf::Spec const &spec, DiagThreadname const &n)
+{
+  bwformat(w, spec, std::string_view{n.name});
+  return w;
+}
+
+} // namespace
+
+
+size_t
+DebugInterface::generate_format_string(swoc::LocalBufferWriter<1024> &format_writer, const char *debug_tag, DiagsLevel diags_level,
+                                       const SourceLocation *loc, DiagsShowLocation show_location, const char *format_string)
+{
+  // Save room for optional newline and terminating NUL bytes.
+  format_writer.restrict(2);
+
+  format_writer.print("[{}] ", DiagTimestamp{});
+  auto timestamp_offset = format_writer.size();
+
+  format_writer.print("{} {}: ", DiagThreadname{}, level_name(diags_level));
+
+  if (location(loc, show_location, diags_level)) {
+    format_writer.print("<{}> ", *loc);
+  }
+
+  if (debug_tag) {
+    format_writer.print("({}) ", debug_tag);
+  }
+
+  format_writer.print("{}", format_string);
+
+  format_writer.restore(2);                  // restore the space for required termination.
+  if (format_writer.view().back() != '\n') { // safe because always some chars in the buffer.
+    format_writer.write('\n');
+  }
+  format_writer.write('\0');
+
+  return timestamp_offset;
 }
